Feat(web-react): Dictionaries constants
